Spmv_struct: fixing bug in matrix construction and adding mv support
There was a bug in the matrix construction, the value of numEntries was not set.
Also adding initial support for multivectors, so it only uses the unstructured spmv.
It also does check for a multivector containing a single vector in which case it calls the vector implementation of spmv_struct.